patched NaN error
There is an error where after fixing the tangled roads, some segments will have one of its vertex having position[Nan, Nan]. This road will not be rendered but still in the road network list.

The root cause of this happening is unknown.

A workaround of using try except is added to patch up the issue. Result appears ok, but a small segment of major road (Highway) disappeared at the area where tangled road happened and fixed.